moved createGardenInDatabase and createUserInDatabase to their own modules
