+++ src/resolvers/Garden.test.ts
@@ -4,6 +4,8 @@ import faker from 'faker'
 import { User } from '../entities/User'
 import { Garden } from '../entities/Garden'
 import jwt from '../utils/jwt'
+import { createGardenInDatabase } from '../utils/test/createGardenInDatabase'
+import { createUserInDatabase } from '../utils/test/createUserInDatabase'
 
 let connection: Connection
 let garden: Garden
@@ -110,25 +112,3 @@ describe('the createGarden mutation', () => {
     })
   })
 })
-
-async function createUserInDatabase(): Promise<User> {
-  const user = User.create({
-    email: faker.internet.email(),
-    password: faker.internet.password(),
-    firstName: faker.name.firstName(),
-    lastName: faker.name.lastName(),
-  })
-  await connection.manager.save(user)
-  return user
-}
-
-async function createGardenInDatabase(
-  owner: User,
-  name: string
-): Promise<Garden> {
-  const garden = Garden.create({
-    name,
-    owner,
-  })
-  return await garden.save()
-}
